move state machine tests into proper mod; use p256k1 from crate exports
